requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
distro==1.9.0
frozenlist==1.8.0
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.18.0
multidict==6.7.0
numpy==2.3.4
oauthlib==3.3.1
openai==2.6.0
openai-agents==0.4.0
packaging==25.0
pluggy==1.6.0
propcache==0.4.1
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.48.0
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wolframclient==1.4.0
yarl==1.22.0
